Rewrite getRouteDescriptions, fix message route not appearing in openapi spec
