+++ src/api/routes/channels/#channel_id/invites.ts
@@ -25,10 +25,10 @@ import {
 	PublicInviteRelation,
 	User,
 	emitEvent,
+	isTextChannel,
 } from "@spacebar/util";
 import { Request, Response, Router } from "express";
 import { HTTPError } from "lambert-server";
-import { isTextChannel } from "./messages";
 
 const router: Router = Router();
