+++ src/api/routes/channels/#channel_id/purge.ts
@@ -25,11 +25,11 @@ import {
 	emitEvent,
 	getPermission,
 	getRights,
+	isTextChannel,
 } from "@spacebar/util";
 import { Request, Response, Router } from "express";
 import { HTTPError } from "lambert-server";
 import { Between, FindManyOptions, FindOperator, Not } from "typeorm";
-import { isTextChannel } from "./messages";
 
 const router: Router = Router();
