+++ src/util/util/index.ts
@@ -41,3 +41,4 @@ export * from "./String";
 export * from "./Token";
 export * from "./TraverseDirectory";
 export * from "./WebAuthn";
+export * from "./Gifs";
